Added new status READY_TO_DELIVER to OPEN DELIVERY
I just received the grocery goods, when the truck drive was drive to our home i see that status was changed to READY_TO_DELIVER. I do not know if that was the 15 minutes time frame. Otherwise I can make an automation to give a signal that the truck driver is on his way.